Content-Based Detection: Basic DOM Parser
JSoup's parse() method has been implemented, which constructs a the Document Object Model (DOM) tree from a given File object. It has been tested so that when the file path is changed, the new HTML file is opened, parsed, and stored in a jsoup.nodes.Document object as a DOM tree.