Content-Based Detection: Detect Title in Top-Half of Page
The code checks if the title is located in the top-half of the page. It does this by looking at all the element nodes under the body tag and finding the current node's location in that list. Then, a percentage is calculated. If the percentage is less than 50%, then the title is in the top-half of the web page.