Content-Based Detection: Finished Title Detection Rules
The last three detection rules for the article title are finished. The rules are checking to see if the user has to page down to see the title, checking that the title's length is between 8 and 50 characters, and checking that the title is not hyper linked. JUnit tests were created for each rule and some new HTML pages were created for testing. One last test was added to go through each test HTML file and see if it contains a title, where 1 does and 6 do not.