Content-Based Detection: Get Publication Date Font Color
For given Element node, it can only be a publication date if the font color is gray, black, or blue. Refactored the title font color detection code so that it can be reused for the publication date detection. It had the same issue as the font size detection that it could not handle multiple lines of style content in the HTML's head.