Content-Based Detection: Finished Author Detection Rules
All of the author detection rules are finished. Rules 1 (font size) and 2 (text length) used the refactored methods from the title and date detection. Rule 3 required a new method to detect if an element's text contains a frequent key word. It was written so that it can be utilized by other detection rules in the future. New test HTML pages were also created.