Content-based Detection: Finished Category Detection Rules
I implemented the 5 category detection rules. The frequent word method was written from scratch because of looking for symbols (|, ->, and >). All other methods was based off the refactored code. There was an issue with my page-down detection that required me to convert the entity tag "&gt;" to ">". This has been fixed. Two new test HTML files have been created for the category detection.