Begin: Link Analysis
Created the basic class and JUnit test class that will hold all of the relevant code to link analysis for the research project. The JUnit test class used for link analysis was added to the class RunResearchTests.java so that all content and URL related tests can be executed together.