Link Analysis: Count Number of Slashes Done
New code to count the number of slashes '/' in the URL. There is no mechanism in place to not count slashes in http:// because all links in the data set start after it. This was used in a new method to determine if the number of slashes is at least 4. If they are 4 or more slashes, then the link should contain an article. Otherwise, it should not.